requirements.txt
streamlit
pandas
matplotlib
openpyxl
python-docx
Pillow